# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bank System API",
        "description": "API documentation for Bank Management System with AI-powered fraud detection. All endpoints require authentication using Laravel Sanctum tokens.",
        "contact": {
            "email": "support@banksystem.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login and get API token (OAuth2 Password Grant)",
                "operationId": "2cc6991997564255fab5a657ec802123",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "grant_type": {
                                        "description": "OAuth2 grant type",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abcdef123456789"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get bank accounts",
                "description": "Returns all accounts for admin, or only user's own accounts for other roles",
                "operationId": "ae8217ff2f08e764dd66574adf89cece",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Savings Account"
                                            },
                                            "number": {
                                                "type": "string",
                                                "example": "1234567890"
                                            },
                                            "available_balance": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 5000
                                            },
                                            "ledger_balance": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 5000
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/accounts/{id}": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get account by ID",
                "description": "Get specific account. Non-admin users can only view their own accounts.",
                "operationId": "42b557925ff956b0eb91a0390d09aa32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Access denied - not your account"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get transactions",
                "description": "Returns all transactions for admin, or only user's own transactions for other roles",
                "operationId": "61f93472ac71805186b9ca85e20583c8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "transaction_code": {
                                                "type": "string",
                                                "example": "TRX123456"
                                            },
                                            "amount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 100
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "credit"
                                            },
                                            "risk_level": {
                                                "type": "string",
                                                "example": "safe"
                                            },
                                            "is_flagged": {
                                                "type": "boolean",
                                                "example": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/transactions/flagged": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get flagged/suspicious transactions (Admin only)",
                "description": "Returns all transactions flagged as suspicious by AI fraud detection system. Admin access required.",
                "operationId": "3b0ac14ff2b40b7555957590ed65d738",
                "parameters": [
                    {
                        "name": "risk_level",
                        "in": "query",
                        "description": "Filter by risk level",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "low",
                                "medium",
                                "high"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_flagged": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "statistics": {
                                                    "properties": {
                                                        "high": {
                                                            "type": "integer",
                                                            "example": 3
                                                        },
                                                        "medium": {
                                                            "type": "integer",
                                                            "example": 7
                                                        },
                                                        "low": {
                                                            "type": "integer",
                                                            "example": 5
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "transactions": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "transaction_code": {
                                                                "type": "string"
                                                            },
                                                            "amount": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "risk_level": {
                                                                "type": "string"
                                                            },
                                                            "analysis_result": {
                                                                "type": "string"
                                                            },
                                                            "is_flagged": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users (Admin only)",
                "operationId": "a2e43da941c44fbd374354bca4c932f2",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "/api/v1/login",
                        "refreshUrl": "/api/v1/refresh",
                        "scopes": {
                            "api": "Access API endpoints"
                        }
                    }
                }
            }
        }
    },
...
}